requirements.txt
# Core
fastapi==0.115.0
uvicorn==0.30.1
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2

# HTTP & Scraping
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.2

# Gemini API
google-generativeai==0.7.2

# Frontend
streamlit==1.38.0

# Utilities
tenacity==8.3.0